Drop Reply command
